refactor: inline storage.ts into LocalStorage adapter files, delete storage.ts
